+++ NEWS.md
@@ -1,7 +1,8 @@
 # outlookMailer 0.1.1
 
-* Bugfix: apparently `Outlook.MailItem` does not have any `From` property.
+* Bugfix: `Outlook.MailItem` does not have any `From` property (it is called `Sender` instead).
 * Add explicit `disconnect_outlook` in tests.
+* Bugfix: signature is not correctly pasted in HTML e-mails. Do not paste on HTML e-mails for now.
 
 # outlookMailer 0.1